dangtm

+++ AppService/TayHo.ProjectManager/PrjectManager.Read/ProjectManager.Read.Api/ViewModels/NS_CongViecDetailResponseViewModel.cs
@@ -9,7 +9,7 @@ namespace ProjectManager.Read.Api.ViewModels
         public int? CongViecId { get; set; }
         public int? GiaiDoanId { get; set; }
         public decimal? DonGia { get; set; }
-        public int? KhoiLuong { get; set; }
+        public decimal? KhoiLuong { get; set; }
         public int? ReasonId { get; set; }
     }
